refactored more components
